+++ src/commands/init.js
@@ -21,12 +21,10 @@ class InitCommand extends TerraformCommand {
     const config = this.getConfigObject();
     const distributor = new Distributor(config);
 
-    return this.getEnvVarsFromAPI().then(data => this.getExtendedProcessEnv(data)).then(() => {
-      return distributor
-        .runActions(['prepare', 'init'], {
-          silent: this.getOption('silent')
-        }).then(() => Promise.resolve('Done'));
-    });
+    return distributor
+      .runActions(['prepare', 'init'], {
+        silent: this.getOption('silent')
+      }).then(() => Promise.resolve('Done'));
   }
 }
