+++ src/commands/plan.js
@@ -21,13 +21,12 @@ class PlanCommand extends TerraformCommand {
   run() {
     const config = this.getConfigObject();
     const distributor = new Distributor(config);
-    return this.getEnvVarsFromAPI().then(data => this.getExtendedProcessEnv(data)).then(() => {
-      return distributor
-        .runActions(['prepare', 'workspaceSelect', 'plan'], {
-          silent: this.getOption('silent'),
-          planDestroy: this.getOption('destroy')
-        }).then(() => Promise.resolve('Done'));
-    });
+
+    return distributor
+      .runActions(['prepare', 'workspaceSelect', 'plan'], {
+        silent: this.getOption('silent'),
+        planDestroy: this.getOption('destroy')
+      }).then(() => Promise.resolve('Done'));
   }
 }
