+++ src/commands/refresh.js
@@ -20,11 +20,10 @@ class RefreshCommand extends TerraformCommand {
   run() {
     const config = this.getConfigObject();
     const distributor = new Distributor(config);
-    return this.getEnvVarsFromAPI().then(data => this.getExtendedProcessEnv(data)).then(() => {
-      return distributor.runActions(['prepare', 'workspaceSelect', 'refresh'], {
-        silent: this.getOption('silent')
-      }).then(() => Promise.resolve('Done'));
-    });
+
+    return distributor.runActions(['prepare', 'workspaceSelect', 'refresh'], {
+      silent: this.getOption('silent')
+    }).then(() => Promise.resolve('Done'));
   }
 }
