+++ src/commands/run.js
@@ -29,17 +29,15 @@ class RunCommand extends TerraformCommand {
       printConfigAsList(this.getConfigObject(), this.getProjectConfig());
       return Promise.resolve('Done');
     }
-    return this.getEnvVarsFromAPI().then(data => this.getExtendedProcessEnv(data)).then(() => {
-      return this._getPromise()
-        .then(answer => {
-          if (answer) {
-            return this._runPhases();
-          } else {
-            return Promise.reject('Action aborted');
-          }
-        })
-        .then(() => Promise.resolve('Done'));
-    });
+    return this._getPromise()
+      .then(answer => {
+        if (answer) {
+          return this._runPhases();
+        } else {
+          return Promise.reject('Action aborted');
+        }
+      })
+      .then(() => Promise.resolve('Done'));
   }
 
   /**
